cleaned up add dataset and delete routes to make them more robust

+++ back_end/app/api/delete_routes.py
@@ -38,7 +38,10 @@ def delete_dataset(dataset_id):
     deletion_queue = [dataset] + intervals + averageIntervalData
     for entry in deletion_queue:
         try:
-            os.remove(entry.file_path)
+            if entry.file_path is not None:
+                os.remove(entry.file_path)
+            else:
+                current_app.logger.warning(f"Tried removing {entry.file_path}, but there was no filepath!")
         except FileNotFoundError:
             current_app.logger.warning(f"Tried removing {entry.file_path}, but file does not exist!")
         db.session.delete(entry) # TODO: this leaves the session invalid for a short time for some reason -> fix!
